+++ epserde/tests/test_types.rs
@@ -371,3 +371,6 @@ fn test_box_box_generic_compiles() -> anyhow::Result<()> {
     assert_eq!(a, full);
     Ok(())
 }
+
+#[derive(Epserde)]
+struct S<A>(A, #[epserde(force_full)] A);